Add explicit type conversions where necessary: gpu/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947888
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891709}

+++ gpu/command_buffer/service/shared_image_backing_d3d.cc
@@ -449,7 +449,8 @@ SharedImageBackingD3D::ProduceDawn(SharedImageManager* manager,
     texture_descriptor.format = wgpu_format;
     texture_descriptor.usage = GetAllowedDawnUsages();
     texture_descriptor.dimension = WGPUTextureDimension_2D;
-    texture_descriptor.size = {size().width(), size().height(), 1};
+    texture_descriptor.size = {static_cast<uint32_t>(size().width()),
+                               static_cast<uint32_t>(size().height()), 1};
     texture_descriptor.mipLevelCount = 1;
     texture_descriptor.sampleCount = 1;
