+++ gpu/ipc/common/gpu_memory_buffer_impl_dxgi.cc
@@ -54,8 +54,8 @@ base::OnceClosure GpuMemoryBufferImplDXGI::AllocateForTesting(
          usage == gfx::BufferUsage::SCANOUT);
 
   D3D11_TEXTURE2D_DESC desc = {
-      size.width(),
-      size.height(),
+      static_cast<UINT>(size.width()),
+      static_cast<UINT>(size.height()),
       1,
       1,
       DXGI_FORMAT_R8G8B8A8_UNORM,
